pyproject.toml
#project
[project]
name = "must-test-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "fastapi[standard]==0.115.0",
    "pydantic-settings>=2.10.1",
]
[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.11",
]

#ruff
[tool.ruff]

target-version = "py312"
extend-exclude = ["migrations"]
line-length = 79
show-fixes = true

[tool.ruff.lint]

select = [
    "UP",   # pyupgrade
    "I",    # isort
    "F",    # flake8-pyflakes
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "PT",   # flake8-pytest-style
    "DTZ",  # flake8-datetimez
]

fixable = [
    "I",
    "F401",
    "T20",
]

[tool.ruff.lint.extend-per-file-ignores]

"tests/*.py" = ["S101", "DTZ005"]

[tool.ruff.format]

quote-style = "double"
indent-style = "space"


#pytest
[tool.pytest.ini_options]

addopts = "-v"
console_output_style = "count"
minversion = "6.0"
pythonpath = ["app", "tests", "."]
testpaths = "./tests"
asyncio_mode = "auto"
filterwarnings = ["ignore::PendingDeprecationWarning"]
